+++ app/src/main/res/layout/fragment_super_hero.xml
@@ -6,7 +6,7 @@
     android:orientation="vertical"
     android:gravity="center"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ui.superhero.SuperHeroFragment">
+    tools:context=".presentation.superhero.SuperHeroFragment">
 
     <androidx.appcompat.widget.SearchView
         android:id="@+id/searchView"
