+++ app/src/main/res/layout/fragment_super_hero_detail.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.superherodetails.SuperHeroDetailFragment"
+    tools:context=".presentation.superherodetails.SuperHeroDetailFragment"
     xmlns:app="http://schemas.android.com/apk/res-auto">
 
   <ImageView
